Fixed bug
Previously, when clicking on the edit icon to
change the ontoClass of an annotation,
it inadvertently opened all the modals for every
annotation. Consequently, users could only interact
 with the modal of the last annotation.

To address this, I implemented a fix that ensures
only the modal corresponding to the clicked
annotation is opened, allowing users to effectively
visualize and modify the ontoClass without
interference from other modals.